server/requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
pydantic
websockets
asyncio
redis
requests
